Added configuration to allow building the project on Circle CI.
This commit also contains a docker image for a Couchbase Server
instance which allows us to run the ottoman tests against a live
Couchbase Server instance rather than only our internal mock.

Extreme special thanks to @Moxious and @erik-etops for putting in
the majority of the work on this commit.  Some unfortunate rebasing
and splitting was needed and thus the author was not preserved.